python_backend/requirements.txt
# Wekeza Multi-Agent Trading System - Python Backend
# Dependencies for deep learning, agents, and trading

# ========== Core ML ==========
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ========== Web Framework ==========
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# ========== Market Data ==========
alpaca-py>=0.10.0
yfinance>=0.2.36
ccxt>=4.2.0

# ========== Multi-Agent Framework ==========
crewai>=0.28.0
langchain>=0.1.0
langchain-groq>=0.0.3

# ========== Vector Database ==========
chromadb>=0.4.22

# ========== Technical Analysis ==========
ta>=0.10.2
mplfinance>=0.12.10b0

# ========== Backtesting ==========
backtesting>=0.3.3
